Rearrange code
Create CodeViewer.tsx for JSON (and later adjacency matrix view). Let SideBar.jsx only be the sidebar, main functionality has been moved to App.tsx. Moved to PromiseWithCancel to interfaces file and directory.
