+++ client/src/Interfaces/PromiseWithCancel.ts
@@ -1,0 +1,5 @@
+export interface PromiseWithCancel<T> extends Promise<T> {
+    cancel: () => void;
+    dateTime: Date;
+    name: string
+}
